+++ src/main/java/pl/plajer/murdermystery/Main.java
@@ -160,6 +160,14 @@ public class Main extends JavaPlugin {
   }
 
   private boolean validateIfPluginShouldStart() {
+    if (!getServer().getPluginManager().isPluginEnabled("CropseReborn") || !getServer().getPluginManager().isPluginEnabled("HolographicDisplays")) {
+      MessageUtils.dependencyWarning();
+      Bukkit.getConsoleSender().sendMessage(ChatColor.RED + "You don't have CorpseReborn or HolographicDisplays installed!");
+      Bukkit.getConsoleSender().sendMessage(ChatColor.RED + "These plugins are required for Murder Mystery to work! Shutting off...");
+      forceDisable = true;
+      getServer().getPluginManager().disablePlugin(this);
+      return false;
+    }
     version = Bukkit.getServer().getClass().getPackage().getName().replace(".", ",").split(",")[3];
     if (!(version.equalsIgnoreCase("v1_11_R1") || version.equalsIgnoreCase("v1_12_R1") || version.equalsIgnoreCase("v1_13_R1")
         || version.equalsIgnoreCase("v1_13_R2"))) {